+++ entities/entity.js
@@ -28,10 +28,10 @@ class Entity extends Sprite {
             onLoadCallback: config.onLoadCallback
         };
         super(options);
-        this.colisionOfsset = config.colisionOfsset || 1
+        this.collisionOffset = config.collisionOffset || 1
         this.hitbox = new Hitbox(new Vector(0, 0), new Vector(this.width, this.height), this.position);
 
-        this.groundSensor = new Hitbox(new Vector(0, this.height ), new Vector(this.width, this.height+1), this.position); // Array to hold ground sensors
+        this.groundSensor = new Hitbox(new Vector(0, this.height ), new Vector(this.width, this.height+this.collisionOffset+0.01), this.position); // Array to hold ground sensors
         this.onGround = false; // Flag to indicate if the entity is on the ground
         this.velocity = new Vector(0, 0); // Velocity vector for the entity
 
@@ -39,7 +39,7 @@ class Entity extends Sprite {
         this.gravity = config.gravity || new Vector(0, 500); // Default gravity if not provided
         this.inAirDrag = config.inAirDrag 
         this.touching = [];
-        this.pasibleOnGround = []; // Flag to indicate if the entity is pasible on ground
+        this.passableOnGround = []; // Flag to indicate if the entity is passable on ground
         config.physic.addEntity(this); // Assuming physic is an instance of a class that manages entities
     }
 
@@ -50,6 +50,7 @@ class Entity extends Sprite {
             //console.log("Entity is on the ground");
             // Update ground sensors positions
             
+            this.onGround = this.passableOnGround.some((a) => a);
         }else{
             this.velocity.add(Vector.mult(this.gravity, deltaTime));
             this.velocity.x *= this.inAirDrag
@@ -57,10 +58,10 @@ class Entity extends Sprite {
         
         // Update the entity's position based on its velocity
         this.velocity.add(Vector.mult(this.acceleration, deltaTime));
-        // if (Math.abs(this.velocity.x) < this.colisionOfsset){
+        // if (Math.abs(this.velocity.x) < this.collisionOffset){
         //     this.velocity.x = 0; // Prevents the entity from moving too slowly
         // }
-        // if (Math.abs(this.velocity.y) < this.colisionOfsset){
+        // if (Math.abs(this.velocity.y) < this.collisionOffset){
         //     this.velocity.y = 0; // Prevents the entity from moving too slowly vertically
         // }
         this.position.add(Vector.mult(this.velocity, deltaTime));
@@ -68,9 +69,8 @@ class Entity extends Sprite {
         // Update the hitbox position
         this.hitbox.updatePosition(this.position);
         this.groundSensor.updatePosition(this.position); // Update the ground sensor position
-        this.onGround = this.pasibleOnGround.some((a) => a);
         
-        this.pasibleOnGround = []; // Reset onGround flag at the start of each update
+        this.passableOnGround = []; // Reset onGround flag at the start of each update
        this.acceleration = new Vector(0, 0); // Reset acceleration after applying it
     }
 
@@ -86,7 +86,7 @@ class Entity extends Sprite {
         this.acceleration.add(acceleration); // Add the given acceleration to the entity's acceleration
     }
 
-    checkColision(block) {
+    checkCollision(block) {
         // Check if the entity's hitbox collides with the block's hitbox
         if (this.hitbox.isColliding(block.hitbox)) {
             this.onCollision(block);
@@ -94,40 +94,40 @@ class Entity extends Sprite {
         }
         if(this.onGround){
             // Check if the entity is touching the ground sensors
-            this.pasibleOnGround.push(this.checkSensor(block));
+            this.passableOnGround.push(this.checkSensor(block));
             
         }
         return false; // No collision
     }
 
     onCollision(block) {
-        let positionDiferenc = Vector.sub(Vector.add(block.position,Vector.div(block.hitbox.offset2,2)), Vector.add(this.position,Vector.div(this.hitbox.offset2,2)));
-        //console.log("v"+positionDiferenc.x+" "+positionDiferenc.y);
-        if (Math.floor(Math.abs(positionDiferenc.x)) > Math.floor(Math.abs(positionDiferenc.y))){
-            if (positionDiferenc.x > 0) {
+        let positionDifference = Vector.sub(Vector.add(block.position,Vector.div(block.hitbox.offset2,2)), Vector.add(this.position,Vector.div(this.hitbox.offset2,2)));
+        //console.log("v"+positionDifference.x+" "+positionDifference.y);
+        if (Math.floor(Math.abs(positionDifference.x)) > Math.floor(Math.abs(positionDifference.y))){
+            if (positionDifference.x > 0) {
                 // Collision from the left
                 //console.log("Collision from the left");
-                this.position.x = block.hitbox.position.x - this.width-this.colisionOfsset;
+                this.position.x = block.hitbox.position.x - this.width-this.collisionOffset;
                 block.onCollision(this, new Vector(1,0)); // Notify the block of the collision
-            } else  if (positionDiferenc.x < 0) {
+            } else  if (positionDifference.x < 0) {
                 //console.log("Collision from the right");
                 // Collision from the right
-                this.position.x = block.hitbox.position.x + block.hitbox.offset2.x+this.colisionOfsset;
+                this.position.x = block.hitbox.position.x + block.hitbox.offset2.x+this.collisionOffset;
                 block.onCollision(this, new Vector(-1,0)); // Notify the block of the collision
             }
         }else {
             
-            if (positionDiferenc.y > 0) {
+            if (positionDifference.y > 0) {
                 // Collision from above
                 //console.log("Collision from above");
-                this.position.y = block.hitbox.position.y - this.height-this.colisionOfsset;
+                this.position.y = block.hitbox.position.y - this.height-this.collisionOffset;
                 block.onCollision(this, new Vector(0,1)); // Notify the block of the collision
                 this.onGround = true; // Set onGround to true
-                this.pasibleOnGround.push(true); // Add to pasibleOnGround array
-            }else  if (positionDiferenc.mag < Math.sqrt(block.height*block.height + block.width*block.width) && positionDiferenc.y < 0) {
+                this.passableOnGround.push(true); // Add to passableOnGround array
+            }else  if (positionDifference.mag < Math.sqrt(block.height*block.height + block.width*block.width) && positionDifference.y < 0) {
                 // Collision from below
                 //console.log("Collision from below");
-                this.position.y = block.hitbox.position.y + block.hitbox.offset2.y+this.colisionOfsset;
+                this.position.y = block.hitbox.position.y + block.hitbox.offset2.y+this.collisionOffset;
                 let velocityY = this.velocity.y
                 block.onCollision(this,new Vector(0,-1)); // Notify the block of the collision
                  // Prevents the entity from sticking to the bottom of blocks when gravity is negative